Use new algorithm to voting

+++ gateway/migrations/2024062123200_MeetUpPaperVotes.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS meet_up_papers_votes (
     paper_id UUID NOT NULL,
     meet_up_id UUID NOT NULL,
     user_id UUID NOT NULL,
-    vote INT NOT NULL,
+    vote DOUBLE NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
     FOREIGN KEY (paper_id) REFERENCES papers(id),
